requirements.txt
absl-py==1.0.0
appdirs==1.4.4
astunparse==1.6.3
autopep8==1.6.0
av==9.0.0
beautifulsoup4==4.10.0
buildozer==1.2.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cycler==0.11.0
Cython==0.29.28
distlib==0.3.4
djitellopy==2.4.0
docutils==0.18.1
easytello==0.0.7
filelock==3.6.0
Flask==2.0.3
flatbuffers==2.0
fonttools==4.29.1
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.2
itsdangerous==2.1.1
Jinja2==3.0.3
keras==2.8.0
Keras-Preprocessing==1.1.2
Kivy==2.1.0
Kivy-Garden==0.1.4
kiwisolver==1.3.2
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.5.1
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
pep517==0.6.0
perfume==0.2
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.1
plyer==2.0.0
protobuf==3.19.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pygame==2.1.2
Pygments==2.11.2
pyler==0.2.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytoml==0.1.21
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sh==1.14.2
six==1.16.0
soupsieve==2.3.1
tello-sim==0.0.1
tellopy==0.6.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
toml==0.10.2
typing-extensions==4.1.1
urllib3==1.26.8
virtualenv==20.13.3
Werkzeug==2.0.3
wrapt==1.14.0
zipp==3.7.0